refactor(api): modernize TensorCast API with Store session management
